msm_mdp4: Fix to build with Toolchain 4.7

+++ drivers/video/msm/mdp4_overlay.c
@@ -2705,7 +2705,7 @@ int mdp4_overlay_play_wait(struct fb_info *info, struct msmfb_overlay_data *req)
 
 	mdp4_mixer_stage_commit(pipe->mixer_num);
 
-	if (mfd->panel_info.pdest == MDP4_MIXER1) {
+	if ((int)mfd->panel_info.pdest == MDP4_MIXER1) {
 		if (dbg_force_ov1_blt)
 			mfd->use_ov1_blt |= (0x1 << 8);
 		else
@@ -2713,7 +2713,7 @@ int mdp4_overlay_play_wait(struct fb_info *info, struct msmfb_overlay_data *req)
 	}
 
 	if (mfd->use_ov1_blt &&
-	    (mfd->panel_info.pdest == MDP4_MIXER1))
+	    ((int)mfd->panel_info.pdest == MDP4_MIXER1))
 		mdp4_overlay1_update_blt_mode(mfd);
 
 	mdp4_overlay_dtv_wait_for_ov(mfd, pipe);
@@ -2857,7 +2857,7 @@ int mdp4_overlay_play(struct fb_info *info, struct msmfb_overlay_data *req)
 		}
 	}
 
-	if (mfd->panel_info.pdest == MDP4_MIXER0) {
+	if ((int)mfd->panel_info.pdest == MDP4_MIXER0) {
 		if (dbg_force_ov0_blt)
 			mfd->use_ov0_blt |= (0x1 << 8);
 		else
@@ -2865,10 +2865,10 @@ int mdp4_overlay_play(struct fb_info *info, struct msmfb_overlay_data *req)
 	}
 
 	if (mfd->use_ov0_blt &&
-	    mfd->panel_info.pdest == MDP4_MIXER0)
+	    (int)mfd->panel_info.pdest == MDP4_MIXER0)
 		mdp4_overlay_update_blt_mode(mfd);
 
-	if (mfd->panel_info.pdest == MDP4_MIXER1) {
+	if ((int)mfd->panel_info.pdest == MDP4_MIXER1) {
 		if (dbg_force_ov1_blt)
 			mfd->use_ov1_blt |= (0x1 << 8);
 		else
@@ -2876,7 +2876,7 @@ int mdp4_overlay_play(struct fb_info *info, struct msmfb_overlay_data *req)
 	}
 
 	if (mfd->use_ov1_blt &&
-	    (mfd->panel_info.pdest == MDP4_MIXER1))
+	    ((int)mfd->panel_info.pdest == MDP4_MIXER1))
 		mdp4_overlay1_update_blt_mode(mfd);
 
 	if (pipe->pipe_type == OVERLAY_TYPE_VIDEO) {
